requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
facenet-pytorch==2.6.0
filelock==3.25.2
fsspec==2026.3.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kagglehub==1.0.0
kagglesdk==0.1.16
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
pillow==12.2.0
protobuf==7.34.1
PyYAML==6.0.3
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
python-dotenv==1.2.2